chore: create standalone imageprovider
